+createSalt ~insertUserBlob ~dbSel ~Tests
Removed any salt or random string creation from the system, and replaced
it with the createSalt method which takes a username and creates a large
random string from it and then hashes it.
Modified the insertUserBlob to return the blob that is inserted so it
can be used both in tests and cookies, since the random string is no
longer being created before inserting a blob and the script which
creates the blob needs to know what the blob is still.
Modified the dbSel method to do searches with things other than
`blank`='blank' but can handle things like less than, or less than or
equal.
Modified many tests to work with these modifications.