~Setup +Example ~addUser ~sanitize ~verification
Modified setup file to create database to contain the tables as well as
added `if not exists` to table creation.
Started the example website (localhost doesn't work: won't send emails,
won't set cookies). Registration works, login works except for setting a
session.
Modified the addUser method in the UserSystem class to check if the
username and email are in use as well as sending an email to the user to
activate, additionally checking for extra data to be added is actually
would work now.
Sanitizing URLs now doesn't automatically fail but works correctly.
Verification of users doesn't default fail if there's no cookie set, but
checks if the necessary cookie is set.